requirements.txt
numpy>=1.24
pandas>=2.0
requests>=2.28
psycopg2-binary>=2.9
pytz>=2023